backend/pyproject.toml
[project]
name = "ai-conversation-backend"
version = "0.1.0"
description = "AI Conversation & Digital Agent System — FastAPI Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "websockets>=12.0",
    "google-genai>=0.8",
    "anthropic>=0.28",
    "google-cloud-texttospeech>=2.16",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "websockets>=12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
